Client handler huge refactor.code quality should be better + it support receiveiving lots of messages then processing it. Not receive->proccess->receive + Now handler recv part and handler execution part runs in separate threads

+++ queue.h
@@ -20,5 +20,5 @@ void* queue_pop(struct queue* drpcq);
 void queue_free_internals(struct queue* drpcq);
 void queue_free(struct queue* drpcq);
 
-uint64_t queue_get_len(struct queue* drpcq);
+size_t queue_get_len(struct queue* drpcq);
